spython_package: fix for venv support
Idempotency was not working when installing venv packages because it
check the status of packages to install in pip attributes, which are
packages installed on the python system, not virtualenvs. The custom
plugin don't retrieve packages installed on virtualenvs, so run
command in virtualenv to check pip freeze result for the package to
install.

Fix spython_runtime for pip & setuptools upgrade: check if `upgrade`
is a boolean with: `!!upgrade == upgrade`, and don't run :upgrade
action if `!upgrade`

Fix spython_venv: don't create virtualenv if already exists.